+++ src/main/resources/schema.sql
@@ -1,3 +1,7 @@
+DROP TABLE IF EXISTS users;
+DROP TABLE IF EXISTS pets;
+DROP TABLE IF EXISTS rooms;
+
 CREATE TABLE IF NOT EXISTS users
 (
     id    BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
@@ -7,6 +11,8 @@ CREATE TABLE IF NOT EXISTS users
     CONSTRAINT pk_user PRIMARY KEY (id),
     CONSTRAINT UQ_USER_EMAIL UNIQUE (email)
 );
+
+INSERT INTO users (name, email, role) values ('boss', 'boss@mail.ru', 'ROLE_BOSS');
 
 CREATE TABLE IF NOT EXISTS pets
 (
@@ -24,8 +30,6 @@ CREATE TABLE IF NOT EXISTS pets
     CONSTRAINT pk_pet PRIMARY KEY (id)
 );
 
--- INSERT INTO users (name, email, role) values ('boss', 'boss@mail.ru', 'ROLE_BOSS');
-
 CREATE TABLE IF NOT EXISTS rooms
 (
     id        BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
@@ -36,4 +40,5 @@ CREATE TABLE IF NOT EXISTS rooms
     available BOOLEAN DEFAULT TRUE,
     CONSTRAINT pk_room PRIMARY KEY (id)
 );
+
