+++ src/test/java/ru/dogudacha/PetHotel/pet/dto/UpdatePetDtoTest.java
@@ -1,37 +1,37 @@
-package ru.dogudacha.PetHotel.user.dto;
+package ru.dogudacha.PetHotel.pet.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import org.springframework.boot.test.json.JsonContent;
-import ru.dogudacha.PetHotel.pet.dto.NewPetDto;
+import ru.dogudacha.PetHotel.pet.dto.UpdatePetDto;
 import ru.dogudacha.PetHotel.pet.model.Sex;
 import ru.dogudacha.PetHotel.pet.model.TypeOfDiet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @JsonTest
-class NewPetDtoTest {
+class UpdatePetDtoTest {
     @Autowired
-    private JacksonTester<NewPetDto> json;
+    private JacksonTester<UpdatePetDto> json;
 
-    final NewPetDto newPetDto = NewPetDto.builder()
-            .typeOfPet("Dog")
-            .breed("Spaniel")
-            .sex(Sex.FEMALE)
-            .age(2)
-            .weight(7)
-            .diet(TypeOfDiet.READY_INDUSTRIAL_FOOD)
-            .isTakesMedications(false)
-            .isContact(true)
-            .isPhotographed(true)
-            .comments("Like play with ball.")
+    final UpdatePetDto updatePetDto = UpdatePetDto.builder()
+            .typeOfPet("Dog small")
+            .breed("Spaniel span")
+            .sex(Sex.MALE)
+            .age(12)
+            .weight(8)
+            .diet(TypeOfDiet.NATURAL_RAW_FOOD)
+            .isTakesMedications(true)
+            .isContact(false)
+            .isPhotographed(false)
+            .comments("Like play with small ball.")
             .build();
 
     @Test
-    void testNewPetDto() throws Exception {
-        JsonContent<NewPetDto> result = json.write(newPetDto);
+    void testUpdatePetDto() throws Exception {
+        JsonContent<UpdatePetDto> result = json.write(updatePetDto);
 
         assertThat(result).extractingJsonPathValue("$.typeOfPet").isNotNull();
         assertThat(result).extractingJsonPathValue("$.breed").isNotNull();
@@ -42,6 +42,5 @@ class NewPetDtoTest {
         assertThat(result).extractingJsonPathValue("$.isTakesMedications").isNotNull();
         assertThat(result).extractingJsonPathValue("$.isContact").isNotNull();
         assertThat(result).extractingJsonPathValue("$.isPhotographed").isNotNull();
-
     }
 }
